// src/models/enum-models/ContentTypeEnum.ts
export enum ContentTypeEnum {
  REEL = "REEL",
  STORY = "STORY",
  POST = "POST",
  TEXT_SHARE = "TEXT_SHARE",
  ARTICLE_SHARE = "ARTICLE_SHARE",
  MEDIA_SHARE = "MEDIA_SHARE",
  TWEET = "TWEET",
  THREAD = "THREAD",
  VIDEO = "VIDEO",
  STATUS = "STATUS",
  TIKTOK_VIDEO = "TIKTOK_VIDEO",
  TIKTOK_IMAGE = "TIKTOK_IMAGE",
  ARTICLE = "ARTICLE",
}